AI

Documentation

data GameTree

The game tree consisting of the board, the current turn and a list of possible moves and the resulting sub-gametree.

Constructors

GameTree

Fields

game_board :: Board

game_turn :: Col

next_moves :: [(Position, GameTree)]

buildTree :: (Board -> Col -> [Position]) -> Board -> Col -> GameTree

A function to generate moves (i.e. board positions) for a player (Col) on a particular board, resulting in a (potentially) infinite game tree. Makes calls to the move generator to construct the tree of positions. relies on lazy evaluation to not construct more nodes than necessary. This means that even though the tree is huge, only the nodes accessed get created.

getBestMove :: Int -> GameTree -> Col -> Position

Get the best next move from a (possibly infinite) game tree. This traverses the game tree up to a certain depth, and pick the move which leads to the position with the best score for the player whose turn it is at the top of the game tree.

type Selector = [(Int, Position)] -> (Int, Position)

Selector type selects a value-move tuple from a list of such tuples.

recurse :: Int -> Selector -> GameTree -> Col -> (Int, Position)

A helper function for finding the best move. Returns a value-move tuple. This recursion goes down the game tree nodes, using the Selector to make predictions about the moves that will be taken. This function changes its selector every time it calls itself. This reflects the presumption that the AI will select the best move and the human player - the worst move (one that puts the AI in the worst position).

switchSelector :: Selector -> Selector

Function swaps the selectors best and worst as every level in the gametree will need a different selector (my best move vs opponents worst move)

best :: Selector

Selector selects the best move from a list of moves and corresponding values.

worst :: Selector

Selector selects the worst move from a list of moves and corresponding values.

greater :: (Int, Position) -> (Int, Position) -> (Int, Position)

Comparator for value-move touples. Returns the greater move.

lesser :: (Int, Position) -> (Int, Position) -> (Int, Position)

Comparator for value-move touples. Returns the lesser move.

aiturn :: World -> World

The AI makes a move by calling getBestMove to determine where to put a piece.

besideFilledCells :: Board -> Col -> [Position]

Given a board will return a list of empty cells beside filled cells

isBesideFilledCell :: Board -> Position -> Bool

Function checks if a given cell is beside a filled cell

emptyCells :: Board -> [Position]

Given a board, this function will return list of positions that are empty This is very slow for large boards. Also this generator of moves is the last if the "besideFilledCells" generator fails.

isEmptyCell :: Board -> Position -> Bool

Given a board and position, function will check if given position is empty

emptyNeighbors :: Board -> Position -> [Position]

Given a board and position, function will return a list of empty neighbor cells

neighbors :: Board -> Position -> [Position]

Function will return all neighbors to a cell (takes care of cells on the edge)

removeDuplicates :: [Position] -> [Position]

Function removes duplicates from a list of positions

hint :: World -> World

Displays a hint based on the ai algorithm in the console for the user.