Board

Documentation

data Col

Colour data - either Black or White

Constructors

Black

White

other :: Col -> Col

Swaps a Col and returns the opposite colour of the one given

type Position = (Int, Int)

Position data type consisting of 2 Integers (Row, Column)

sq_border :: Color

Constants like sizes, line-colours, margins, etc

sq_side :: Float

bar_side1 :: Float

bar_side2 :: Float

bar_text_scale :: Float

bar_margin :: Float

win_side1 :: Float

win_side2 :: Float

win_text_scale :: Float

win_margin :: Float

win_size :: Int -> Int

wwh :: Fractional a => Int -> a

data Net_Data

Record for Network data stored inside the world record.

Constructors

Net_Data

Fields

useNet :: Bool

isServ :: Bool

socket :: Maybe Socket

addr :: String

port :: String

initNet_Data :: Net_Data

default data (addr of localhost)

data Board

A Board record containing the board size (a board is a square grid, n * n), the number of pieces in a row required to win, and a list of pairs of position and the colour at that position.

Constructors

Board

Fields

size :: Int

target :: Int

human :: Col

pieces :: [(Position, Col)]

won :: Maybe Col

fair :: Bool

initBoard :: Board

Default board is 6x6, target is 3 in a row, no initial pieces and not fair

data World

Overall state is the board and whose turn it is, plus any further information about the world (this may later include, for example, player names, timers, information about rule variants, etc)

Constructors

World

Fields

board :: Board

turn :: Col

cmd :: String

ai_on :: Bool

replay :: Maybe String

recording :: Maybe String

is_menu :: Bool

blacks :: Picture

whites :: Picture

cell :: Picture

prev :: Maybe World

net_data :: Net_Data

pic :: World -> Col -> Picture

Function returns the Gloss Picture for blacks or whites from the workd

initWorld :: World

Defualt initial world using initBoard and initNet_Data

outOfBounds :: Board -> Position -> Bool

checks if a position is out of bounds of the board

makeMove :: Board -> Col -> Position -> Maybe Board

Play a move on the board; return Nothing if the move is invalid (e.g. outside the range of the board, or there is a piece already there or the move breaks the rules)

eliminate :: Maybe a -> a

Function eliminates a Maybe

checkWon :: Board -> Maybe Col

Check whether the board is in a winning state for either player. Returns Nothing if neither player has won yet Returns 'Just c' if the player c has won

checkRules :: Board -> Bool

returns true if board follows rules, false if rule is broken

check4and4 :: Board -> Col -> Int -> Bool

returns true if board follows rules, false if rule is broken

check3and3 :: Board -> Col -> Int -> Bool

returns true if board follows rules, false if rule is broken

data Direction

Direction data record

Constructors

N

S

E

W

NE

SE

NW

SW

next :: Position -> Direction -> Position

Funcion returns the next cell given a direction and starting cell

opposite :: Direction -> Direction

Flips the directions

opp :: Direction -> Direction

colOf :: Board -> Position -> Maybe Col

Retruns the colour of a position on the board, or Nothing if empty

descend :: Board -> Col -> Direction -> Position -> (Int, Bool)

Continues along the board in a direction along a line of pieces on the board

longest :: Board -> Col -> Int

Finds the longest row of pieces of 1 colour that are not blocked by enemy pieces or are a victory row. Descends from every square into all directions.

evaluate :: Board -> Col -> Int

An evaluation function for a minimax search. Given a board and a colour returns an integer indicating how good the board is for that colour.

save :: FilePath -> World -> IO World

Saves the world to a given file path.

load :: World -> FilePath -> IO World

Loads a world from a specified filepath. Keeps the images from the given world but changes the other components to what was loaded