GameWorld

Documentation

data Col

Colour data - either Black or White

Constructors

Black

White

other :: Col -> Col

Swaps a Col and returns the opposite colour of the one given

type Position = (Int, Int)

Position data type consisting of 2 Integers (Row, Column)

sq_border :: Color

Constants like sizes, line-colours, margins, etc

sq_side :: Float

win_side1 :: Float

win_side2 :: Float

win_text_scale :: Float

win_margin :: Float

win_size :: Int -> Int

wwh :: Fractional a => Int -> a

data Net_Data

Record for Network data stored inside the world record.

Constructors

Net_Data

Fields

useNet :: Bool

isServ :: Bool

socket :: Maybe Socket

addr :: String

port :: String

initNet_Data :: Net_Data

default data (addr of localhost)

data Board

A Board record containing the board size (a board is a square grid, n * n), the number of pieces in a row required to win, and a list of pairs of position and the colour at that position.

Constructors

Board

Fields

size :: Int

target :: Int

human :: Col

pieces :: [(Position, Col)]

won :: Maybe Col

fair :: Bool

initBoard :: Board

Default board is 6x6, target is 3 in a row, no initial pieces and not fair

data World

Overall state is the board and whose turn it is, plus any further information about the world (this may later include, for example, player names, timers, information about rule variants, etc)

Constructors

World

Fields

board :: Board

curr_menu :: Menu

turn :: Col

cmd :: String

ai_on :: Bool

ai_level :: Int

replay :: Maybe String

recording :: Maybe String

is_menu :: Bool

to_network :: Bool

taking_add :: Bool

taking_port :: Bool

blacks :: Picture

whites :: Picture

cell :: Picture

prev :: Maybe World

net_data :: Net_Data

pic :: World -> Col -> Picture

Function returns the Gloss Picture for blacks or whites from the workd

initWorld :: World

Defualt initial world using initBoard and initNet_Data

outOfBounds :: Board -> Position -> Bool

checks if a position is out of bounds of the board

makeMove :: Board -> Col -> Position -> Maybe Board

Play a move on the board; return Nothing if the move is invalid (e.g. outside the range of the board, or there is a piece already there or the move breaks the rules)

eliminate :: Maybe a -> a

Function eliminates a Maybe

checkWon :: Board -> Maybe Col

Check whether the board is in a winning state for either player. Returns Nothing if neither player has won yet Returns 'Just c' if the player c has won

checkRules :: Board -> Bool

returns true if board follows rules, false if rule is broken

checkRule :: Board -> Col -> Int -> Bool

returns true if board follows a given rule, false if rule is broken

type Street = (Position, Direction, (Int, Bool))

Street type used in 3and3 and 4and4. Stores the result of a descend + extra info. Touple storing the Position, the Direction of the row, and the result of the check (Length of row, if row is blocked).

data Direction

Direction data record

Constructors

N

S

E

W

NE

SE

NW

SW

next :: Position -> Direction -> Position

Funcion returns the next cell given a direction and starting cell

opposite :: Direction -> Direction

Flips the directions

opp :: Direction -> Direction

colOf :: Board -> Position -> Maybe Col

Retruns the colour of a position on the board, or Nothing if empty

descend :: Board -> Col -> Direction -> Position -> (Int, Bool)

Continues along the board in a direction along a line of pieces on the board

findEnd

Arguments

:: Board

-> Col

-> Direction

-> Position
last position

-> Position
current position to check

-> Position
end of row

Finds the end of a row of pieces on the board

longest :: Board -> Col -> Int

Finds the longest row of pieces of 1 colour that are not blocked by enemy pieces or are a victory row. Descends from every square into all directions.

evaluate :: Board -> Col -> Int

An evaluation function for a minimax search. Given a board and a colour returns an integer indicating how good the board is for that colour.

undo :: World -> World

Takes a world and, if playing against an AI will return world of player's previous move, if playing another player it will return the world of the previous turn, if there is no previous world it will return last possible world

save :: FilePath -> World -> IO World

Saves the world to a given file path.

load :: World -> FilePath -> IO World

Loads a world from a specified filepath. Keeps the images from the given world but changes the other components to what was loaded

data Menu

Menu data containing MenuEntries and decorations (unclickable Gloss pictures like lines or titles)

Constructors

Menu

Fields

entries :: [MenuEntry]

options :: [MenuEntryOption]

decorations :: [MenuEntryUnclickable]

data MenuEntry

Constructors

MenuEntry

Fields

menu_draw :: Picture

func :: World -> World

location :: Point

data MenuEntryOption

Constructors

MenuEntryOption

Fields

option_draw :: Picture

option_func :: World -> World

option_name :: String

option_location :: Point

var :: Bool

data MenuEntryUnclickable

Constructors

MenuEntryUnclickable

Fields

decoration_draw :: Picture

menuClick :: Point -> World -> Maybe World

handleOption :: World -> MenuEntryOption -> World

Handles function of option and updates option in menu to display current state

flipMenuOption :: MenuEntryOption -> MenuEntryOption

Changes display of option to opposite value of current value

replaceMenuOption :: Menu -> MenuEntryOption -> Menu

Taking Menu and MenuEntryOption will return

singlePlayerOptions :: World -> World

localPlayChoice :: World -> World

setThreeAndThree :: World -> World

Sets game rules to include the rules of three and three and four and four

setAI :: World -> World

setHardAI :: World -> World

multiPlayerChoiceLocal :: World -> World

Given a world will return world ready to take an address and port for connecting to a host

multiPlayerChoiceHost :: World -> World

Given a world will return world ready to take an port for becoming a host

multiPlayerChoiceJoin :: World -> World

Given a world will return world ready to take an address and port for connecting to a host

setToHost :: World -> World

Given a world will prepare it for becoming a host

setToJoin :: World -> World

Given a world will prepare it for connecting to a host

takePort :: World -> String -> World

Given a World and a String representing a port will return given world prepared for hosting networking

takeAdd :: World -> String -> Maybe World

Take World and string as representative of ip address and port, returns world if prepared for connecting networking if network/port is valid else returns nothing

submitAddressAndPort :: World -> World

Takes World where command is representative of ip address and port, returns world if prepared for networking if network/port is valid else returns current world

submitPort :: World -> World

Takes current command stored in the world as a port, returns world with port variable and other variables set for networking

resetWorld :: World -> World

Resets world to initial world, keeps images for pieces and board

initMenu :: Menu

optionMenu :: Menu

connectJoinChoiceMenu :: Menu

connectHostChoiceMenu :: Menu

mainMenuTitle :: MenuEntryUnclickable

localEntry :: MenuEntry

multiPlayerHostEntry :: MenuEntry

multiPlayerJoinEntry :: MenuEntry

submitTextEntry :: MenuEntry

getPortTextDecoration :: MenuEntryUnclickable

submitTextEntry2 :: MenuEntry

getPortTextDecoration2 :: MenuEntryUnclickable

localOptionsTitle :: MenuEntryUnclickable

localPlayBegin :: MenuEntry

fairOption :: MenuEntryOption

aiOption :: MenuEntryOption

aiHardOption :: MenuEntryOption

cancelEntry :: MenuEntry

changeMenu :: World -> Menu -> World

Takes world and menu, returns world with current menu set to given menu

bar_side1 :: Float

MenuEntry data containing dimensions for drawing menu boxes and taking input

bar_side2 :: Float

bar_text_scale :: Float

bar_margin :: Float

title_side1 :: Float

title_side2 :: Float

title_text_scale :: Float

menu_box_R :: Float

menu_box_G :: Float

menu_box_B :: Float

menu_box_A :: Float

isInBounds :: Point -> Menu -> Maybe MenuEntry

Given a point and a menu will return menuEntry, in menu, point is in, if it is in one, otherwise returns nothing

isInBoundsOptions :: Point -> Menu -> Maybe MenuEntryOption

Given a point and menu will return a MenuEntryOption if the given point is inside the boundaries of any picture of the menu

isInBar :: Point -> MenuEntry -> Bool

Given a menu entry will return boolean representing if the point is in the menu entry's location in the gui

isInBarOption :: Point -> MenuEntryOption -> Bool

Given a point and MenuEntryOption will return Boolean representing whether the point is in the picture for the MenuEntryOption

inRangeX :: Float -> Float -> Bool

Given two floats representing x positionswith one representing x position of point and other representing x position of menuBar will return if the value is in the menu bar's x coordinates

inRangeY :: Float -> Float -> Bool

Given two floats representing y positions with one representing x position of point and other representing x position of menuBar will return if the value is in the menu bar's x coordinates

menuBar :: Point -> String -> Picture

Given a point and a string generates a menu bar picture containing given text

menuBox :: Point -> Picture

Given a point will return

menuText :: Point -> String -> Picture

Given a point and String will produce a picture of the text as would fit within a text bar

optionBar :: Point -> String -> Bool -> Picture

Given a point, text and a value representing an option location, name and value will produce a picture representing an option and it's current value at that point

optionStatusDisplay :: Point -> Bool -> Picture

Given a point and Bool will return a box with a text value representing the boolean value at given point

optionStatusBox :: Point -> Picture

Given a point will produce a picture of an box at that point

optionStatusText :: Point -> Bool -> Picture

Given a point and a value will return a text picture of that given value as on or off (True or False)

titleBar :: Point -> String -> Picture

Given a point and a string generates a title bar picture containing given text

titleBox :: Point -> Picture

Given a point will return

titleText :: Point -> String -> Picture

Given a point and String will produce a picture of the text as would fit within a text bar