Input

Documentation

handleInput :: Event -> World -> World

Function updates the world state given an input event. Mouse-click : converts the x,y coordinates to the corresponding row,col and calls makeMove with the world, colour and coordinates. If makeMove returns Nothing, the move was invalid and the world is not updated. If a new world is returned then the move was valid and it is sent over the network if necessary. Also switches the turn. KeyboardiInpput : "Ctrl+z" undos the 2 last moves (own move and opponent) "Ctrl+s" saves the game to a file 'game.dat' "Ctrl+l" loades a game from a file 'game.dat' _ other commands unimplimented.