Main

Documentation

parseArgument :: String -> World -> World

Function parses an argument and modifies the world accordingly if the argument is recognized. Ignores undrecognised arguments. Please consult the README.md for further details.

main :: IO ()

Main method sets up world using any program arguments, sets up networking if needed, loads bmp pictures and configures the Gloss play function.