Network

Documentation

posToString :: (Int, Int) -> String

Converts a (x,y) position to a String to send over the network "(x,y)"

stringToPos :: String -> (Int, Int)

Reads a String into a (x,y) position.

sendMove :: World -> (Int, Int) -> IO World

Sends a move over a connected socket. Converts the position to string first.

sendMsg :: Socket -> String -> IO Int

Sends a given string over a connected socket.

rcvMsg :: Socket -> String

Waits on an incoming message from a given Socket. Returns when a string has been recieved. Limit is currently 1024 bytes which is more than enough for th strings we are sending.

rcvMove :: World -> Socket -> World

Uses rcvMsg to recieve a string from an connected socket and converts it into a (x,y) position. Then adds the position with the turn's colour to the pieces of the board. Note: Does not check for validity of move. This is the responsibility of the other socket endpoint. For this practical we assume that we are only connected to other instances of our game.

setupNetworking :: World -> IO World

unsafeSetupNetworking :: World -> World

Runs setupNetworking discarding IO() for when the game is started with networking from the Menu

serverSetup :: String -> IO Socket

Sets up the server and waits on an incoming connection when the socket is ready. Returns a successfully connected Socket. Fails if port is already in use or permissions are not sufficient.

clientSetup :: String -> String -> IO Socket

Sets up the client and attempts to connect to an endpoint with the specified address and port. Fails if endpoint doesnt exist or refuses connection.