Draw

Documentation

drawWorld :: World -> Picture

Given a world state, return a Picture which will render the world state. This will need to extract the Board from the world state and draw it as a grid plus pieces.

grid :: World -> Picture

Draws the grid as an 2d array of squares

square :: Point -> World -> Picture

Draws a square by using the bmp picture and translating it to the right coordinate

drawMenu :: World -> Picture

Draws the menu using the Gloss Pictures specified in the world.

drawCurrentCmd :: World -> Picture

Returns a box with the current cmd of the world given drawn inside it

tiles :: World -> Picture

Draws all the board squares (tiles). There should be size*size of them.

tile :: World -> (Position, Col) -> Picture

Draws a single board square (tile). Uses the gloss picture supplied in world and translates to the right position

winBox :: Point -> Picture

Draws the win-message-box if a player has won.

winmsg :: Int -> Maybe Col -> Picture

Draws the whole win-message if a player has won.

winText :: Point -> String -> Picture