Recording

Documentation

readAll :: String -> String

Reads a file into a string.

encodeMove :: Col -> Position -> String

Encodes a given move into a string.

decodeMove :: String -> (Position, Col)

Decodes a String into a move (Position, Col).

decode :: String -> [(Position, Col)]

Decodes a string into a list of moves. Uses decodeMove until the whole string is parsed.

encodeSize :: Int -> String

Function converts a board size (given as an int) into a string.

decodeSize :: String -> Int

Function parses a string into a board size.

data Record

Contains information about parsed recording. Size of the board, a list of the move-color tuples, and how many moves have been read from the list.

Constructors

History

Fields

bsize :: Int

moves :: [(Position, Col)]

moves_read :: Int

initRecord :: String -> Record

The initial record to start from. Given a path to file, decode the size and then the moves. Start with 0 mvoes read.

nextmove :: Record -> ((Position, Col), Record)

Returns the next move and increments the number of moves read.

buildSequence :: Record -> World -> World

Function builds worlds until a victory state is reached.

sequenceStart :: World -> World

Starts the buildSequence to replay a game from a filepath.

writeMove :: String -> (Position, Col) -> t -> IO t

Writes a given move to the file specifed and returns IO (second arg)

wrmv :: World -> (Position, Col) -> t -> t

Wraps writeMove so IO is not spread.

writeSize :: String -> World -> IO World

Writes the size of a world to the file and returns IO(World)

wrsz :: World -> World

Wraps writeSize so IO is not spread.